requirements.txt
discord.py
aiohttp
discord-logger
get-stick-bugged-lol
requests